@@ -15,7 +15,7 @@
   },
   "dependencies": {
     "dayjs": "1.11.10",
-    "next": "13.4.19",
+    "next": "14.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "5.0.1",
@@ -24,18 +24,19 @@
   "devDependencies": {
     "@babel/core": "7.23.9",
     "@next/eslint-plugin-next": "13.5.6",
-    "@storybook/addon-actions": "7.0.27",
-    "@storybook/addon-essentials": "7.0.27",
-    "@storybook/addon-interactions": "7.0.27",
-    "@storybook/addon-links": "7.0.27",
-    "@storybook/nextjs": "7.0.27",
-    "@storybook/react": "7.0.27",
+    "@storybook/addon-actions": "7.6.10",
+    "@storybook/addon-essentials": "7.6.10",
+    "@storybook/addon-interactions": "7.6.10",
+    "@storybook/addon-links": "7.6.10",
+    "@storybook/nextjs": "7.6.10",
+    "@storybook/react": "7.6.10",
     "@storybook/testing-library": "0.2.2",
     "@types/node": "20.11.7",
     "@types/react": "18.2.48",
     "@typescript-eslint/eslint-plugin": "6.19.1",
     "@typescript-eslint/parser": "6.19.1",
     "eslint": "8.56.0",
+    "eslint-config-next": "14.1.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-import-resolver-typescript": "3.6.1",
     "eslint-plugin-import": "2.29.1",
@@ -44,8 +45,12 @@
     "prettier": "3.2.4",
     "sass": "1.70.0",
     "serve": "14.2.1",
-    "storybook": "7.0.27",
+    "storybook": "7.6.10",
     "typescript": "5.3.3"
+  },
+  "resolutions": {
+    "string-width": "4.2.2",
+    "strip-ansi": "6.0.0"
   },
   "browserslist": [
     "defaults"
