@@ -1,3 +1,5 @@
+'use client'
+
 import { Footer } from '@/components/Footer'
 import { Navigation } from '@/components/Navigation'
 import { NewsList } from '@/components/Newslist'
@@ -5,7 +7,7 @@ import { SectionTitle } from '@/components/SectionTitle'
 import { WidthContainer } from '@/components/WidthContainer'
 import { articles, talks, interviews } from '@/data/news'
 
-import styles from './news.module.scss'
+import styles from './page.module.scss'
 
 const NewsPage = () => (
   <div className={`root ${styles.root}`}>
